requirements.txt
blinker==1.9.0
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.0
Flask==3.1.1
fonttools==4.58.5
imbalanced-learn==0.13.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lightgbm==4.6.0
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.3
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pluggy==1.6.0
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.5.1
scipy==1.15.3
seaborn==0.13.2
shap==0.48.0
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
threadpoolctl==3.6.0
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
Werkzeug==3.1.3
xgboost==3.0.2
